requirements.txt
Django==1.8.9
django-leaflet==0.18.0
psycopg2==2.6.1
factory-boy==2.6.0
django-bootstrap-form==3.2
